Extracte ImagePage widget to its own file and removed some unused imports
