+++ parka/lib/components/images/parka_image_card_widget.dart
@@ -6,6 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:get/get.dart';
 
+import 'image_page.dart';
+
 class ParkaImageCardWidget extends StatelessWidget {
   final PlaceHolderType type;
   final Function onTapHandler;
@@ -93,7 +95,7 @@ class ParkaImageCardWidget extends StatelessWidget {
                           ),
                         ),
                       )
-                    : CarImagePLaceholder(
+                    : CarImagePlaceholder(
                         cardWidth: cardWidth,
                         type: this.type,
                       ),
@@ -102,59 +104,6 @@ class ParkaImageCardWidget extends StatelessWidget {
           },
         ),
       ),
-    );
-  }
-}
-
-class ImagePage extends StatelessWidget {
-  final String image;
-  ImagePage(this.image);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.black,
-      ),
-      backgroundColor: Colors.black,
-      body: Hero(
-          tag: this.image, child: Center(child: Image.network(this.image))),
-    );
-  }
-}
-
-class CarImagePLaceholder extends StatelessWidget {
-  final PlaceHolderType type;
-
-  CarImagePLaceholder({
-    Key key,
-    @required this.cardWidth,
-    @required this.type,
-  }) : super(key: key);
-
-  final double cardWidth;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        SvgPicture.asset(
-          this.type == PlaceHolderType.Car
-              ? "resources/images/carPlaceHolder.svg"
-              : "resources/images/parkingPlaceHolder.svg",
-          height: cardWidth / 6,
-          width: cardWidth / 6,
-        ),
-        Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Icon(
-            Icons.add,
-            color: Colors.white,
-            size: cardWidth / 6,
-          ),
-        ),
-      ],
     );
   }
 }
