+++ parka/lib/components/modals/parking_detail_modal.dart
@@ -1,5 +1,5 @@
 import 'package:ParkA/components/buttons/circle_button.dart';
-import 'package:ParkA/components/images/parka_image_card_widget.dart';
+import 'package:ParkA/components/images/image_page.dart';
 import 'package:ParkA/components/place-holders/no_image_placeholder.dart';
 import 'package:ParkA/components/rating/star_rating.dart';
 import 'package:ParkA/controllers/user_controller.dart';
